+++ gui/src/components/Settings/Wallet/HDWallet.tsx
@@ -21,25 +21,30 @@ import { Address, formatUnits } from "viem";
 import { z } from "zod";
 
 import { useProvider } from "@/hooks";
-import {
-  derivationPathSchema,
-  hdWalletSchema,
-  passwordFormSchema,
-  Wallet,
-} from "@/types";
+import { passwordFormSchema, passwordSchema } from "@/types/password";
+import { derivationPathSchema, mnemonicSchema, Wallet } from "@/types/wallets";
 
-const createSchema = hdWalletSchema.extend({
+export const schema = z.object({
+  count: z.number().int().min(1).max(100),
+  name: z.string().min(1),
+  current: z.array(z.string()).length(2).optional(),
+  mnemonic: mnemonicSchema,
+  derivationPath: derivationPathSchema,
+  password: passwordSchema,
+});
+
+const createSchema = schema.extend({
   current: z.string(),
 });
 
-const updateSchema = hdWalletSchema.pick({
-  type: true,
+const updateSchema = schema.pick({
   name: true,
   derivationPath: true,
   count: true,
 });
 
-type FormType = z.infer<typeof createSchema> | z.infer<typeof updateSchema>;
+type CreateSchema = z.infer<typeof createSchema>;
+type UpdateSchema = z.infer<typeof updateSchema>;
 
 const steps = ["Import", "Secure", "Review"];
 
@@ -47,23 +52,19 @@ interface Props {
   type: "create" | "update";
   wallet?: Wallet & { type: "HDWallet" };
 
-  onSubmit: (data: FormType) => void;
+  onSubmit: (data: CreateSchema | UpdateSchema) => void;
   onRemove: () => void;
 }
 
-export function HDWalletForm({ type, ...props }: Props) {
-  if (type === "create") {
-    return <HDWalletCreateForm {...props} />;
+export function HDWalletForm({ wallet, ...props }: Props) {
+  if (!wallet) {
+    return <Create {...props} />;
   } else {
-    return <HDWalletUpdateForm {...props} />;
+    return <Update wallet={wallet} {...props} />;
   }
 }
 
-function HDWalletCreateForm({
-  wallet,
-  onSubmit,
-  onRemove,
-}: Omit<Props, "type">) {
+function Create({ onSubmit, onRemove }: Omit<Props, "type">) {
   const [name, setName] = useState("");
   const [step, setStep] = useState(0);
   const [mnemonic, setMnemonic] = useState<string>("");
@@ -75,7 +76,6 @@ function HDWalletCreateForm({
   useEffect(() => {
     if (!current || !mnemonic || !derivationPath || submitted) return;
     onSubmit({
-      type: "HDWallet",
       count: 5,
       name,
       mnemonic,
@@ -84,16 +84,7 @@ function HDWalletCreateForm({
       password,
     });
     setSubmitted(true);
-  }, [
-    name,
-    current,
-    mnemonic,
-    derivationPath,
-    password,
-    wallet,
-    onSubmit,
-    submitted,
-  ]);
+  }, [name, current, mnemonic, derivationPath, password, onSubmit, submitted]);
 
   return (
     <Stack direction="column" spacing={2}>
@@ -385,11 +376,7 @@ function NativeBalance({ address }: NativeBalanceProps) {
   );
 }
 
-function HDWalletUpdateForm({
-  wallet,
-  onSubmit,
-  onRemove,
-}: Omit<Props, "type">) {
+function Update({ wallet, onSubmit, onRemove }: Omit<Props, "type">) {
   const {
     register,
     handleSubmit,
