+++ gui/src/components/Settings/Wallet/Impersonator.tsx
@@ -1,13 +1,13 @@
 import { zodResolver } from "@hookform/resolvers/zod";
 import { Button, Stack, TextField } from "@mui/material";
+import { Address } from "abitype";
 import { useFieldArray, useForm } from "react-hook-form";
 import { z } from "zod";
 
-import { addressSchema, ImpersonatorWallet } from "@/types";
+import { addressSchema, ImpersonatorWallet, Wallet } from "@/types/wallets";
 
 // react-hook-form doesn't support value-arrays, only object-arrays, so we need this type as a workaround for the impersonator form
 export const schema = z.object({
-  type: z.literal("impersonator"),
   name: z.string().min(1),
   addresses: z.array(
     z.object({
@@ -21,7 +21,7 @@ type Schema = z.infer<typeof schema>;
 
 export interface Props {
   wallet?: ImpersonatorWallet;
-  onSubmit: (data: ImpersonatorWallet) => void;
+  onSubmit: (data: Wallet) => void;
   onRemove: () => void;
 }
 
@@ -49,7 +49,8 @@ export function ImpersonatorForm({ wallet, onSubmit, onRemove }: Props) {
   const prepareAndSubmit = (data: Schema) => {
     onSubmit({
       ...data,
-      addresses: data.addresses.map(({ address }) => address),
+      type: "impersonator",
+      addresses: data.addresses.map(({ address }) => address as Address),
     });
     reset(data);
   };
