--- gui/src/types.ts
@@ -1,221 +1,0 @@
-import { invoke } from "@tauri-apps/api/tauri";
-import { Address } from "viem";
-import { z } from "zod";
-import { zxcvbn } from "zxcvbn-typescript";
-
-// const formSchema = schema.shape.network;
-export const networkSchema = z.object({
-  networks: z.array(
-    z
-      .object({
-        name: z.string().min(1),
-        explorer_url: z.string().optional().nullable(),
-        http_url: z.string().min(1),
-        ws_url: z.string().nullable().optional(),
-        currency: z.string().min(1),
-        chain_id: z.number(),
-        decimals: z.number(),
-      })
-      .refine(
-        (data) =>
-          data.chain_id !== 31337 || (!!data.ws_url && data.ws_url.length > 0),
-        {
-          path: ["ws_url"],
-          message: "WebSockets are mandatory for dev networks",
-        },
-      ),
-  ),
-});
-
-export const passwordSchema = z.string().superRefine((password, ctx) => {
-  const { feedback, score } = zxcvbn(password);
-
-  if (score < 4) {
-    const message =
-      feedback.suggestions.length > 0
-        ? feedback.suggestions.join(" ")
-        : "Please use a stronger password.";
-
-    ctx.addIssue({
-      code: z.ZodIssueCode.custom,
-      message,
-    });
-  }
-});
-
-export const passwordFormSchema = z
-  .object({
-    password: passwordSchema,
-    passwordConfirmation: z.string(),
-  })
-  .refine((data) => data.password === data.passwordConfirmation, {
-    path: ["passwordConfirmation"],
-    message: "The two passwords don't match",
-  });
-
-export const mnemonicSchema = z
-  .string()
-  .regex(/^([a-z]+\s)+[a-z]+$/, {
-    message: "Must be a list of english words",
-  })
-  .refine(
-    (data) => {
-      const words = data.split(/\s+/).length;
-      return [12, 15, 18, 21, 24].includes(words);
-    },
-    {
-      message:
-        "Invalid number of words. Needs to be 12, 15, 18, 21 or 24 words long",
-    },
-  )
-  .refine(
-    (mnemonic) => invoke<string>("wallets_validate_mnemonic", { mnemonic }),
-    {
-      message: "Invalid mnemonic. You may have a typo or an unsupported word",
-    },
-  );
-
-export const derivationPathSchema = z
-  .string()
-  .regex(/^m\/(\d+'?\/)+\d+$/, {
-    message: "invalid path format",
-  })
-  .default("m/44'/60'/0'/0");
-
-export const addressSchema = z
-  .string()
-  .refine((data) => data.match(/^0x[a-fA-F0-9]{40}$/), {
-    message: "not a valid ETH address",
-  });
-
-// react-hook-form doesn't support value-arrays, only object-arrays, so we need this type as a workaround for the impersonator form
-export const addressOrObjectSchema = z.union([
-  addressSchema,
-  z.object({
-    addressSchema,
-  }),
-]);
-
-export const hdWalletSchema = z.object({
-  type: z.literal("HDWallet"),
-  count: z.number().int().min(1).max(100),
-  name: z.string().min(1),
-  current: z.array(z.string()).length(2).optional(),
-  mnemonic: mnemonicSchema,
-  derivationPath: derivationPathSchema,
-  password: passwordSchema,
-});
-
-export const hdWalletUpdateSchema = hdWalletSchema.pick({
-  type: true,
-  name: true,
-  derivationPath: true,
-  count: true,
-});
-
-export const jsonKeystoreSchema = z.object({
-  type: z.literal("jsonKeystore"),
-  name: z.string().min(1),
-  file: z.string().min(1),
-  currentPath: z.string().optional(),
-});
-
-export const plaintextSchema = z.object({
-  type: z.literal("plaintext"),
-  name: z.string().min(1),
-  mnemonic: mnemonicSchema,
-  derivationPath: derivationPathSchema,
-  count: z.number().int().min(1),
-  currentPath: z.string().optional(),
-});
-
-export const impersonatorSchema = z.object({
-  type: z.literal("impersonator"),
-  name: z.string().min(1),
-  addresses: z.array(addressSchema).min(1),
-  current: z.number().optional(),
-});
-
-export const walletSchema = z.discriminatedUnion("type", [
-  hdWalletSchema,
-  jsonKeystoreSchema,
-  plaintextSchema,
-  impersonatorSchema,
-]);
-
-export const walletTypes: Wallet["type"][] = Array.from(
-  walletSchema.optionsMap.keys(),
-)
-  .filter((k) => !!k)
-  .map((k) => k as unknown as Wallet["type"]);
-
-export const walletsSchema = z.object({
-  wallets: z.array(walletSchema),
-});
-
-export type Wallet = z.infer<typeof walletSchema>;
-export type HdWallet = z.infer<typeof hdWalletSchema>;
-export type JsonKeystoreWallet = z.infer<typeof jsonKeystoreSchema>;
-export type PlaintextWallet = z.infer<typeof plaintextSchema>;
-export type ImpersonatorWallet = z.infer<typeof impersonatorSchema>;
-
-export type Network = z.infer<typeof networkSchema.shape.networks>[number];
-
-export interface TokenBalance {
-  contract: Address;
-  balance: string;
-  metadata: TokenMetadata;
-}
-
-export interface TokenMetadata {
-  name: string;
-  symbol: string;
-  decimals: number;
-}
-
-export interface Tx {
-  hash: `0x${string}`;
-  from: Address;
-  to: Address;
-  value: string;
-  data: string;
-  blockNumber: number;
-  position: number;
-  status: number;
-}
-
-export interface Peer {
-  origin: string;
-  tab_id?: number;
-  title?: string;
-  socket: string;
-  url: string;
-  favicon: string;
-}
-
-export interface Pagination {
-  page?: number;
-  page_size?: number;
-}
-
-export interface Paginated<T> {
-  pagination: Pagination;
-  items: T[];
-  last: boolean;
-  total: number;
-}
-
-export type Affinity = { sticky: number } | "global" | "unset";
-
-export const generalSettingsSchema = z.object({
-  darkMode: z.enum(["auto", "dark", "light"]),
-  abiWatch: z.boolean(),
-  abiWatchPath: z.string().optional().nullable(),
-  alchemyApiKey: z.string().optional().nullable(),
-  etherscanApiKey: z.string().optional().nullable(),
-  hideEmptyTokens: z.boolean(),
-  onboarded: z.boolean(),
-  fastMode: z.boolean(),
-});
-
-export type GeneralSettings = z.infer<typeof generalSettingsSchema>;
