+++ gui/src/types/network.ts
@@ -1,0 +1,27 @@
+import { z } from "zod";
+
+// const formSchema = schema.shape.network;
+export const networkSchema = z.object({
+  networks: z.array(
+    z
+      .object({
+        name: z.string().min(1),
+        explorer_url: z.string().optional().nullable(),
+        http_url: z.string().min(1),
+        ws_url: z.string().nullable().optional(),
+        currency: z.string().min(1),
+        chain_id: z.number(),
+        decimals: z.number(),
+      })
+      .refine(
+        (data) =>
+          data.chain_id !== 31337 || (!!data.ws_url && data.ws_url.length > 0),
+        {
+          path: ["ws_url"],
+          message: "WebSockets are mandatory for dev networks",
+        },
+      ),
+  ),
+});
+
+export type Network = z.infer<typeof networkSchema.shape.networks>[number];
