+++ gui/src/types/wallets.ts
@@ -1,0 +1,99 @@
+import { invoke } from "@tauri-apps/api";
+import { z } from "zod";
+
+import { passwordSchema } from "./password";
+
+export const mnemonicSchema = z
+  .string()
+  .regex(/^([a-z]+\s)+[a-z]+$/, {
+    message: "Must be a list of english words",
+  })
+  .refine(
+    (data) => {
+      const words = data.split(/\s+/).length;
+      return [12, 15, 18, 21, 24].includes(words);
+    },
+    {
+      message:
+        "Invalid number of words. Needs to be 12, 15, 18, 21 or 24 words long",
+    },
+  )
+  .refine(
+    (mnemonic) => invoke<string>("wallets_validate_mnemonic", { mnemonic }),
+    {
+      message: "Invalid mnemonic. You may have a typo or an unsupported word",
+    },
+  );
+
+export const derivationPathSchema = z
+  .string()
+  .regex(/^m\/(\d+'?\/)+\d+$/, {
+    message: "invalid path format",
+  })
+  .default("m/44'/60'/0'/0");
+
+export const addressSchema = z
+  .string()
+  .refine((data) => data.match(/^0x[a-fA-F0-9]{40}$/), {
+    message: "not a valid ETH address",
+  });
+
+// react-hook-form doesn't support value-arrays, only object-arrays, so we need this type as a workaround for the impersonator form
+export const addressOrObjectSchema = z.union([
+  addressSchema,
+  z.object({
+    addressSchema,
+  }),
+]);
+
+export const hdWalletSchema = z.object({
+  type: z.literal("HDWallet"),
+  count: z.number().int().min(1).max(100),
+  name: z.string().min(1),
+  current: z.array(z.string()).length(2).optional(),
+  mnemonic: mnemonicSchema,
+  derivationPath: derivationPathSchema,
+  password: passwordSchema,
+});
+
+export const jsonKeystoreSchema = z.object({
+  type: z.literal("jsonKeystore"),
+  name: z.string().min(1),
+  file: z.string().min(1),
+  currentPath: z.string().optional(),
+});
+
+export const plaintextSchema = z.object({
+  type: z.literal("plaintext"),
+  name: z.string().min(1),
+  mnemonic: mnemonicSchema,
+  derivationPath: derivationPathSchema,
+  count: z.number().int().min(1),
+  currentPath: z.string().optional(),
+});
+
+export const impersonatorSchema = z.object({
+  type: z.literal("impersonator"),
+  name: z.string().min(1),
+  addresses: z.array(addressSchema).min(1),
+  current: z.number().optional(),
+});
+
+export const walletSchema = z.discriminatedUnion("type", [
+  hdWalletSchema,
+  jsonKeystoreSchema,
+  plaintextSchema,
+  impersonatorSchema,
+]);
+
+export const walletTypes: Wallet["type"][] = Array.from(
+  walletSchema.optionsMap.keys(),
+)
+  .filter((k) => !!k)
+  .map((k) => k as unknown as Wallet["type"]);
+
+export type Wallet = z.infer<typeof walletSchema>;
+export type HdWallet = z.infer<typeof hdWalletSchema>;
+export type JsonKeystoreWallet = z.infer<typeof jsonKeystoreSchema>;
+export type PlaintextWallet = z.infer<typeof plaintextSchema>;
+export type ImpersonatorWallet = z.infer<typeof impersonatorSchema>;
